Change matching procedure (as discussed with Dominik and Peter on 26th November 2012):
The RGE solver does not run the model to the matching scale.  It
simply calls the match_*() functions.  Within these functions the one
has to make sure that both models are matched at the correct scale
(via calling run_to() or setScale() etc.).

This has the advantage that, in case of a dynamic matching scale, the
matching condition class can take care that the models are correctly
matched.