Change the running algorithm a bit:
In order to make the convergence tester work correctly, all models
must be run to their lowest constraint / matching condition scale.
Until now, this was not done for the lowest model, but we need to do
this for the above reason!

However, the constraint of the lowest model should not be applied
twice (on run_down() and run_up()).  Therefore run_down() only runs
the lowest model down, without applying the constraint.

After the iteration has ended the lowest constraint is applied for a
last time.