Remove interface function Constraint<Two_scale>::update_scale() because because we will not force the two scale solver to call first apply() and then update_scale().
Reason: At the M_susy constraint in the MSSM we first have to 1.
calculate the DRbar parameters, 2. recalculate the M_susy scale and
3. do the EWSB.  I.e. the scale has to be updated between the two
calculations.  This was not possible in the old version, because there
we first do a calculation and then update the scale.

Implications: Now every constraint is responsible for updating the
scale.  This can be done within the apply() function.