Relax deviation in fermion self-energies.
Peter Athron wrote:  _Softsusy corrections_

tau self energy:
Softsusy uses the full bpmz self-energy (ie finite part of the self
energy as defined in e.g. eqn 53 of our threshold paper) *minus* the
photon contribution.  The photon contribution is removed because this
is already accounted for in the SM MS-bar mass that the corrections
are applied to.  There is also a conversion to DR-bar (see eqn 3.5 of
the softsusy manual)

bottom self energy:
As above full BPMZ minus photon correction and with a shift to DR-bar
(eqn 3.3 in the manual). The mb corrections are also resummed as we've
discussed previously.

top self energy:
Softsusy uses the full BPMZ corrections *plus* the some extra two loop
qcd bits (and some further parts he sometimes adds and used in a
recent paper on the focus point region).  Unlike the others for the
top it is the pole mass to which the corrections are applied tp get
the DRbar mass.