Bugfix: We can't use the one-loop corrected Mu and BMu parameters to calculate the tree-level masses, because then the goldstone boson masses will not be equal to the vector boson masses.  Therefore, everytime we call calculate_DRbar_parameters() we have to ensure that the tree-level EWSB eqs. are satisfied.
However, if we want to calculate the one-loop pole masses, we must
take the one-loop corrected Mu and BMu.  Therefore we call
solve_ewsb_one_loop() directly before calculate_1loop_masses().  The
mass matrices which are used inside calculate_1loop_masses() will then
be calculated with the one-loop corrected Mu and BMu.  This is how
SPheno does it.

So, roughly speaking

  1. tree-level Mu, BMu  =>  tree-level spectrum

  2. tree-level spectrum + one-loop EWSB => one-loop spectrum