Add test for the low-scale MSSM (lowMSSM)
The test works like this: We generate a CMSSM spectrum, which we write
to a SLHA file (the output scale is the susy scale).  This point is
then used as input to the lowMSSM.  The low-scale specrum which the
lowMSSM puts out should be the same as the MSSM one.