Pass the tuple processor via const reference for performance reasons.
However, I don't understand why this works, because the tuple
porcessor is the result of std::bind.  The function parameter which is
bound, is a reference to a class, which will be modified by the
function call (see for example SLHA_io::process_modsel_tuple).  But
how can this object be modified when the tuple processor is passed as
const reference?