Bugfix: avoid InverseFunction::ifun errors in the NMSSM
In the NMSSM the InverseFunction::ifun error appears if we try to
solve the EWSB equations for Kappa.  This is because the EWSB
eqs. contain Kappa and T[Kappa], which confuses Mathematica (it can't
solve T[Kappa] for Kappa).

Therefore we have to replace all relevant soft parameters by unique
symbols.  In the NMSSM where we use {Kappa, vS, mS2} as output we
replaces in the EWSB eqs. T[Kappa] -> TKappa.  Afterwards we call
Solve[] and substitute Kappa -> T[Kappa] in the solution.