Bugfix: correct association between tree-level and one-loop tadpoles
In the MRSSM the CP even Higgses are defined as

  {{phid, phiu, phiS, phiT}, {hh, ZH}} .

However, the VEVs are ordered like

  {vd, vu, vT, vS} ,

i.e. the order of the vevs is different from the order of the
corresponding Higgs component fields in hh.  So, we have to ensure
that in the four tadpole equations we correctly associate

  vd   <-->   Tadpole of hh(1)
  vu   <-->   Tadpole of hh(2)
  vT   <-->   Tadpole of hh(4)
  vS   <-->   Tadpole of hh(3)

Thanks to Philip Diessner.