do not change the semantics of ThetaStep
Recently ThetaStep[i,k] was expanded to i < k, because it was assumed
that k is always an integer literal.  But there is no guarantee that
this is always the case.  Therefore, we should always expand

  ThetaStep[i,k]  -->  i <= k

As a result, we have to decrease index literals within ThetaStep in
order to convert to C index convention (just as we do in Delta).