allways output in SLHA format
This simplifies the output logic does not confuse the user in case
no SLHA output file is given.

Note: We should add the MODSEL, MINPAR and EXTPAR block to the output
SLHA if they are not given as input (for example if no SLHA input file
is given or the default values are used).