SARAH workaround: Replace MatMul[] by ScalarProd[] in the traces
The trace abbreviations (RGETraces.m) contain expressions of the form

   MatMul[V, M, V]

where V is a vector and M is a matrix.  Since the traces have to be
scalars we transform these expressions to

   ScalarProd[V, MatMul[M, V]]

which is ensured to yield a scalar.  However, we assume here that the
MatMul[] expressison which appear in the traces are always of the from

   MatMul[vector, ..., another-vector]

If this is not the case, than we have a problem.