Make sure all temporary latex files are removed on clean-doc
We must avoid the usage of $(DIR) in makefile rules, because variables
in makefile rules are resolved at run-time.  But the DIR variable is
overwritten by modules that are loaded after doc.  Therefore we have
to put the list of temporary latex files into a separate variable,
which will not be overwritten.