Bugfix: do not hard-code running Weinberg angle
For the calculation of the (tree-level) DR-bar masses we need the
running DR-bar Weinberg angle, expressed in terms of the fundamental
model parameters (gauge couplings, vevs, ...).  This commit uses
SARAH's expressions for the Weinberg angle (ThetaW), Mass[VWm] and
Mass[VZ] to express ThetaW in terms of these fundamental parameters.