disable Solve::ifun warning when calculating the Weinberg angle
The equation that we're trying to solve contains Sin[] and Cos[] of
the Weinberg angle.  Mathematica's Solve[] function will use the
inverse functions of Sin[] and Cos[] to solve the eq.  But since Sin[]
and Cos[] are invertible only on a finite interval, some solutions
will therefore not be found.  Therefore this warning will appear.

However, we're restricting us to the case where all gauge couplings
are positive and real, so some solutions are irrelevant anyway.
Furthermore we will restrict ourselfs to one solution.  Therefore it
is save to suppress this warning.