try the SM definition of the Weinberg angle first
This increases the probability that we do not need to call Solve[]
because the SM definition is correct in most models.  However, all of
this will fail in the UMSSM where the Higgs quantum numbers are not
fixed (=> there will only be a solution if the quantum numbers are
fixed to some non-anomalous values).