run `make distclean' in standalone/ sub-directory
Since the Makefile in the standalone/ sub-directory is not integrated
into the makefile module tree, we need to cd into the standalone/
sub-directory and run `make distclean' separately.

Note: standalone/Makefile is not included in the makefile module tree
by intention!  We want to provide a completely independent sub-package
with it's own Makefile.