Do not hard code internal color index range
ResolveColorFactor[] in Vertices.m obtains the range of an internal
color index from SARAH`Particles.  For a color octet running in the
loop for instance, the range now becomes from 1 to 8 instead of from 1
to 3 which used to be hard coded.  However, this introduces a
discrepancy between FlexiSUSY and SARAH in the coupling Cp[USd[{gO1}],
conj[USd[{gO2}]], conj[SOc], SOc] which is generated in MRSSM.  The
reason is that a color index in SARAH`sumOverNonAbelianIndizes[] is
always from 1 to 3, returned by SARAH`getIndexRange[].  This issue
needs to be resolved.