adding two-loop dMA contributions to the CP-even Higgs self-energy
Note: dMA contains also two-loop tadpole contributions, which we
cancel out.

On 08.03.2014 02:40, Peter Athron wrote:

Please see http://arxiv.org/pdf/hep-ph/0105096v2.pdf Eqns 3-9. Here
you see how the double derrivatives of the psedouscalar states can
appear in the expressions of the cp even self energies.

Now look at eqn 24.  here is the *definition* of DeltaM_S^2. Slavich
presents his results as elements of DeltaM_S^2 and this is what you
get from his twoloophiggs.f.  The same is true for the yukawa
corrections in http://arxiv.org/pdf/hep-ph/0305127v3.pdf (see eqns
6-10) and the corresponding routines in twoloophiggs.f. He is *not*
simply giving double derivatives or self energies.

The essential point is that he has written his expressions such that
they are to be added on to expressions including the pole mA.  However
as we discussed yesterday Ben uses the DR-bar mAsq in the CP even self
energy and so he must correct for this.  that is what the dMA pieces
do.

Now Ben only adds one loop tadpoles.  This should be because when he
adds the dMA pieces, they are not just the double derivatives of the
pseudoscalar states which we want to cancel(in e.g eqn 10 of
hep-ph/0305127) , but also include the tadpole corrections. dMA is the
two loop corrections to take the DRbar psuedoscalar mass ( =
-2m3sq/sin(2beta)) to the pole mass (as in eqn 8 of hep-ph/0305127)
and thus contains two loop tadpoles.

Therefore if you simply use our current approach where the twoloop
tadpoles are included in the T_i/v_i but add the two loop corrections
and dMA pieces as in softsusy then I believe you will be double
counting the tadpoles!  But you also cannot simply ignore the dMA
pieces and just uses DeltaMS^2 as you would then be including double
derivatives of psuedoscalar states!