correcting self-energy contribution to CP-odd MSSM Higgs
The dMA pieces need to be added to the self-energy matrix in the form
of Eq. (9) hep-ph/0105096 .  dMA contains tadpole contributions (in
form of bA = sb^2 t1/v1 + cb^2 t2/v2, see hep-ph/9606211), which we
need to subtract.

On 10.03.2014 04:42, Peter Athron wrote:

Ben's expression for poleMasq is just BPMZ eqn E6 with the two loop
added on.  Now E6 is the most overly convoluted expression in the
history of the universe.  To get this they start off with the DR-bar
expression for mAsq after EWSB substitution, ie -m3sq/(sb*cb) and add
the one loop self energy, PiAA and the tadpoles, bA (this is what we
need to know).  They then substitute for the DRbar part using first
the one loop EWSB equations (E1 and E2 with T1 and T2 appropriately
set) and then substitute for the musq which appears using the one loop
expression for musq in E6.

The important point though is all they are really doing is adding
Pi_AA and bA to the DRbar mass.  bA takes the form it does because of
the rotation from the gauge basis to the mass eigenstates where there
is only one non-zero mass.

At two loop dMA correction contains both the two loop bA (rotated
tadpoles) and the two loop PiAA self energy (or rotated Pi_P_iP_j if
you prefer).

Now if you look at http://arxiv.org/pdf/hep-ph/0105096v2.pdf equation
9 I think you see everything you need to go from dMA to dMP_ij, ie dMA
inside MA^2 should be mulitipled with the appropriate factor factors
to get the the correction to the MP^2 elements.  Since dMA^(2loop)
includes bA^(2loop) you need to subtract this straight from dMA,
before multiplying with the approriate factors for the individual
elemnents.

Or alternatively to correct for the tapoles after adding the dMA
pieces in:

  MP_11 you want to subtract
      bA^(2loop) * sb^2
         = sb^2 * (sb^2 t_1^(2loop)/v_1 + cb^2 t_2^(2loop)/v_2)

  MP_22 you want to subtract
      bA^(2loop) * cb^2
         = cb^2 * (sb^2 t_1^(2loop)/v_1 + cb^2 t_2^(2loop)/v_2)

  MP_12 you want to subtract
      bA^(2loop) * sb * cb
         = sb*cb * (sb^2 t_1^(2loop)/v_1 + cb^2 t_2^(2loop)/v_2)