Bugfix: generalize color summation routine
1. FS now selects color indices (not) to contract, not by the forms of
the associated generation indices (which are not available if there is
only a single generation), but by comparing each field inside
SARAH`Cp[] with f where f is from SelfEnergies`FSSelfEnergy[f, __].
This enables color summation for instance of Cp[SOc, conj[SOc],
conj[SOc], SOc].

2. FS now sums over one color index common to the two fields being
contracted (as SARAH 4.1.0 does).  For contraction of color triplets,
this leads to the same result as the previous implementation which did
double summation over two independent color indices.  For color
octets, the results differ.  The new implementation makes more sense.
The result still needs to be checked.