Revert "Simplify config/remove_export_markers.sh"
This reverts commit 995b884f4f641c9bccafff8749fd058ecbf32e17.
--
However, you removed a subtile feature: When calling

   ./config/remove_export_markers.sh

(i.e. without any arguments) the script will stuck.  In the previous
version I catched this this by the following if condition:

   if [ ! -t 0 ]; then
      # do not wait for more stdin
   fi

Would you mind adding such a condition again, please?  The reason is
that this script is called inside make like this

  ./config/remove_export_markers.sh $(SOME_VARIABLE)

And in the stupid case where $(SOME_VARIABLE) is empty (for what
reason what so ever), make will stuck, because the script stucks.

Sorry, if I'm too pedantic here.  :)

Best,
Alex