Bugfix: impose EWSB before calculating the spectrum
The last steps in the spectrum generator are to run to the susy scale
and calculate the pole mass spectrum.  However, the running to the
susy scale might spoil the EWSB.  Therefore, solve_ewsb() has to be
called directly after the run to the susy scale and before we
calculate the spectrum.