make EWSB root finder configurable
This configuration option is a sectet at the moment and used for
testing only.  Currently, we're trying the GSL routines "Hybrid",
"HybridS" and "Broyden".  The Newton method might find additional
roots, but is much slower.

Note: BOOST supports root finding as well.  We should try this out at
some point.