Relax MSSM-lowMSSM test due to the varying 1st and 2nd Yukawa couplings
The 1st and 2nd generation quark Yukawa couplings are relatively
sensitive to the low-energy quark masses (SMINPUTS).  When the SG
recalculates these quark masses the result differs by up to 1.7 %
compared to the original input.  This deviation is transferred to the
SLHA input file for the lowMSSM, which results in a 1.7 % deviation of
the lowMSSM Yukawa couplings compared to the the MSSM Yukawa
couplings.

Here is the original test output:

Error: difference between test/MSSM.out.spc and test/lowMSSM.out.spc larger that 3.9e-4
----------------
@ Absolute error = 1.3395330000e-8, Relative error = 1.8307667013e-3
----------------
@ Absolute error = 8.6498900000e-6, Relative error = 2.5865757499e-3
----------------
@ Absolute error = 2.5681900000e-7, Relative error = 1.8229027535e-3
----------------
@ Absolute error = 5.6230100000e-6, Relative error = 1.8229098993e-3
----------------
@ Absolute error = 4.5528490000e-7, Relative error = 1.6560246730e-2
----------------
@ Absolute error = 9.4138860000e-5, Relative error = 1.6560291540e-2