calculate DR-bar MZ and MW at 1-loop order only if thresholds enabled
At the low-scale constraint MW and MZ (in the DR-bar scheme) are used
to calculate the Weinberg angle.  The Weinberg angle is used to
calculate g1 and g2.  I.e. MW and MZ are used to calculate threshold
corrections to g1 and g2.  Therefore, the calculation of DR-bar MW and
MZ from pole MW and MZ at the one-loop level should be done only if
threshold corrections are enabled.  If threshold corrections are
disabled, one should set MW(DR-bar) = MW(pole) and MZ(DR-bar) =
MZ(pole) .