Correcting definition of CKM matrix in terms of Vd and Vu
In FlexibleSUSY the SVD of the SM Fermion mass matrix `mass_matrix_Fd'
(= YD^T) is defined as:

  fs_svd(mass_matrix_Fd, MFd, ZDL, ZDR);

where

  mass_matrix_Fd == ZDL.transpose() * MFd.matrix().asDiagonal() * ZDR

i.e.

  MFd.matrix() == ZDL.conj() * mass_matrix_Fd * ZDR.adjoint()

In Softsusy and SLHA-2 the fermion masses are defined as:

  displayYukawaMatrix(YD).diagonalise(Vd, Ud, yd);

where

  yd = Vd^T * YD * Ud

     = Ud^T * YD^T * Vd   (Eq. (8) in SLHA-2)

This means, that we have the correspondance

  ZDL.conj()     ==  Vd^T
  ZDR.adjoint()  ==  Ud

Note, that SARAH's/FlexibleSUSY's Yukawa couplings are defined to be
transposed compared to SLHA-2.
On 18.10.2014 17:28, jhyeon wrote:
> Hi Alex, if I look at the generated MSSM code, `V_u` and `V_d` are
> initialized to `ZUL` and `ZDL`, respectively, which I think are the
> hermitian conjugates of V_u and V_d appearing in Eq. (8) of SLHA2.