help FS finding the soft Higgs masses
I introduces the new user variable FSSoftHiggsMasses where the user
can specify the list of soft Higgs masses.

For the calculation of the Higgs masses, the tree-level EWSB
eqs. substitution must be done in the Higgs mass matrix.  In FS this
substitution is done like this:

 0) save the soft Higgs masses in temporary varibles

 1) solve the tree-level EWSB eqs. for the soft Higgs masses
    (thereby the soft Higgs masse are overwritten)

 2) calculate the Higgs mass matrix

 3) reset the soft Higgs masses to the saved values from 0)

For these steps, FS needs to know which model parameters are the soft
Higgs masses.  In some models (for example in the munuSSM), this is
difficult to find out.  If FS fails to find the soft Higgs masses, the
user can assist by setting them in the variable FSSoftHiggsMasses.