Bugfix: better guess for momentum used in LowPrecision
In LowPrecision we use as momentum in the self-energies

  p = m_i * m_j          (1)

where m_i is the ith tree-level (DR-bar) mass eigenvalue.  However,
the tree-level mass eigenvalues are mass ordered, which leads to a bad
momentum guess in cases where the mass matrix has the form

  M = ( m11 .   )        (2)
      ( .   m22 )

with m11 >> m22.  To obtain a better momentum guess, the values m_i,
which go into the calculation of the momentum guess (1) should
correspond to the diagonal elements of the tree-level mass matrix (2).

This fix reorders the tree-level masses, which go into (1), according
to the diagonal elements of (2).