copy calculated W pole mass from low-scale constraint to model class
Note: If the low-energy constraint uses G_Fermi as input, the W pole
mass is predicted and stored in the QedQcd object in the low-energy
constraint.  This patch copies the predicted value to the model class,
thereby overwriting the value calculated by the model class.

If the low-energy constraint uses MW as input, the W pole mass is not
predicted.  This patch then copies the input value of MW to the model
class, thereby overwriting the value calculated by the model class.