Revert "Simplify calculation of top pole mass 2-loop calculation"
This reverts commit e29a0545c795a7f5d573c94bbe1f37454fa7bd54.

This revert is necessary in order to not spoil the 2-loop QCD top mass
corrections in HighPrecision mode.

Rationale: The 2-loop contributions assume that the momentum is set to
the dr-bar top mass, p = mt_drbar .  However, in HighPrecision mode
there is an iteration over p = mt_pole .  Therefore, in order to
correctly use the 2-loop contributions (which account for the correct
momentum) the 1-loop contributions have to be calculated for p =
mt_drbar.  Therefore, we strip them out of the self-energies and add
them at the end for p = mt_drbar, together with the 2-loop pieces
(which are also calculated for p = mt_drbar).

Conflicts:
	meta/LoopMasses.m