remove input parameter struct member from the constraints
Reason: The input parameter struct is a member of the model class.
The reaons for this is, that charges might in principle be input (in
the UMSSM for example) or the sign of mu.  However, we should avoid
redundancy and therefore remove the additional input parameter struct
from all constraints.