correcting mass dimension of finite 1-loop tadpole terms
In MS-bar renormalized models finite 1-loop terms in the tadpoles and
self-energies arise.  These terms can be proportional to a mass or a
squared mass.  In SARAH these terms are marked as x*Mass[p] or
x*Mass2[p], where x is a number and p is a particle name.

Before this commit the finite Mass2[p_] terms in the tadpoles were
replaced by M[p], which is wrong.  The only replacement that can be
done is Mass[p_] -> M[p].  This removes the replacement Mass2[p_] ->
M[p], which results in properly finite squared mass terms in the
tadpoles.