chop beta function values smaller than zero_threshold
Reason: Certain beta-functions can be numerically non-zero due to the
imprecision of numerical operations, even if the beta-function should
be algebraically zero.  Since such tiny non-zero beta-functions
sometimes fluctuate, which can confuse the Runge-Kutta integrator,
leading to non-perturbative running.  An example of such a behaviour
can be found in the CMSSMCPV for values of eta != 0.

In order to avoid such problems, a zero-threshold was introduced in
the previous commit.  Beta-functions smaller than this zero-threshold
are set to zero using the Chop() functor.