treat points with non-perturbative parameters as invalid
Now, we flag an error if a non-perturbative dimensionless parameter
occurs.  Before, we printed only a warning.

Reason: The NonPerturbative exception, which is thrown from the
Runge-Kutta integrator, is not a sufficient indicator for a
non-perturbative point.  (The Runge-Kutta indicator throws, only if
too many steps occur, or parameter values become non-finite.)

With the current approach, points with non-perturbative dimensionless
parameters at the last iteration step are marked as unphysical.  The
occurrence of non-perturbative parameters in intermediate steps is
ignored, as long as the non-perturbativity is not present in the last
step of the iteration.