Bugfix: ignore goldstone boson pole mass tachyons
In this commit we skip the calculation of the goldstone boson "pole
mass", because it is gauge dependent and therefore unphysical.  In
this way goldstone boson pole mass tachyons are automatically ignored.
The mass of the goldstone boson in the `physical' struct is set to the
tree-level mass of the corresponding vector boson.