Bugfix: decrease assumed goldstone boson tree-level mass precision
We skip the calculation of the goldstone boson pole masses.  To
determine which mass eigenstate is a goldstone boson, we compare the
tree-level Higgs/Goldstone boson mass with the corresponding vector
boson mass.  This comparison can fail, if the precision of the
goldstone boson tree-level mass is poor.  This commit relaxes the
assumed goldstone boson tree-level mass precision to 1e-8.