allow the user to chose the scale where the EWSB conditions are solved
Fixing the EWSB output parameters at some scale is regarded as a
boundary condition.  The scale where the EWSB output parameters shall
be fixed can be selected via the FSSolveEWSBFor[{...}] symbol.  If
FSSolveEWSBFor[{...}] is not given in any constraint, it will be added
automatically to the susy-scale constraint.

Example:  fixing mu2 at the low-scale

EWSBOutputParameters = { mu2 };

LowScaleInput = {
   {v, 2 MZDRbar / Sqrt[GUTNormalization[g1]^2 g1^2 + g2^2]},
   {Yu, Automatic},
   {Yd, Automatic},
   {Ye, Automatic},
   FSSolveEWSBFor[EWSBOutputParameters]
};