if pole masses of SUSY particles were not set, use tree-level masses
Background: The user might omit to set SUSY pole masses in the MASS
block, and instead set only Mu, M1, M2 etc.  In this case, we
calculate the tree-level mass spectrum using these parameters, and
interpret them as "pole masses".  In this way, the user-defined
parameters Mu, M1, M2 etc. are automatically interpreted as on-shell
parameters.  This again allows one to directly input on-shell
parameters.