Bugfix: use smuon mixing matrix which diagonalizes mass matrix
Before this commit, the smuon "pole mass" matrix was used, which is
the mixing matrix that diagonalizes the loop-corrected smuon mass
matrix for p = m_smuon1.

Now we're using the mixing matrix, which diagonalizes the "tree-level"
mass matrix.  The tree-level mass matrix contains the soft mass
parameters, which have been determined such that the tree-level smuon
mass = the pole mass.