Bugfix: DR-bar -> OS conversion in limit tan(beta) -> infinity
Before this commit, the conversion ml2(1,1) and me2(1,1) from DR-bar
to on-shell did not converge in the limit tan(beta) -> infinity (for
example tan(beta) = 100000).  The reason was that in the smuon mass
matrix the "tree-level" Yukawa coupling (defined in terms of the pole
masses, w/o tan(beta) resummation) was used.

This commit implements an initial guess for the resummed Yukawa
couplings *before* the DR-bar to on-shell conversion for ml2(1,1) and
me2(1,1) is performed.  In this way, the resummed Yukawa coupling is
used in the calculation of the smuon mass matrix during the DR-bar to
on-shell conversion.

Author: Dominik Stockinger