Bugfix: use user defined value of Z pole mass for low-energy scale
Before this commit, the hard-coded value of MZ = 91.1876 GeV was
always used as scale for the low-energy constraint.  This commit
allows the user to use the value of MZ given in the SLHA input file as
low-energy scale.

Note: The low-energy scale is still fixed to be MZ.  It cannot be set
to a scale != MZ.  The reason for this is, that the current
implementation of the calculation of the Weinberg angle at the
low-energy scale assumes that the renormalization scale is MZ.