let CreateLocalConstRefs[] handle DependenceNums
If an expression contain a symbol (not a function!) that has a
DependenceNum, the CreateLocalConstRefs[] function creates a local
const reference for the symbol.

For example for ThetaW in the expression

   g1 ThetaW + g2 ThetaW

a local const ref will be created.  However, for the expression

   g1 ThetaW[] + g2 ThetaW[]

no local const ref will be created, because ThetaW[] is a function.