Bugfix: do multiplet check on fields, non on EWSB conditions
This makes the SSM work correctly.

In the SSM, the 1st EWSB condition and 1st tadpole are non-zero,
because they correspond to the Higgs.  The 2nd EWSB condition and
tadpole should be zero, because the correspond to the singlet field,
which does not develop a VEV.

Thanks to John McDowall and Peter Athron