enable calculation of weiberg angle at an arbitrary scale
The scale dependence in the W and Z self-energies is correctly taken
into account.  The for the calculation of the vertex and box diagrams
in the SM, the scale-dependence had to be added explicitly.  This
scale dependence was obtained from Eq. (6.15) of Lisa Zeune's diploma
thesis

  https://inspirehep.net/record/1235362/

as follows:

First, I have checked, that (6.15) is equivalent to (C.12) from BPMZ
for the case Q = MZ.

In (6.15) only the 1st term on the r.h.s. contains the scale
dependence.  This term is contained in (C.12) for the case Q = MZ.

=>

To account for the scale dependence in (C.12), this term must be (i)
subtracted for Q = MZ, and (ii) added again for arbitrary Q.  This
leads to the expression

   term = -((alphaHat*Log[MW^2/Q^2])/(Pi*sHat^2));

   (term /. Q -> MZ) - term

=>

   -((alphaHat*4*Log[MZ^2/Q^2])/(4*Pi*sHat^2))

I added this term to delta_VB_SM.