disable check of 3L soft-squared beta functions against Softsusy
The 3-loop soft-squared mass beta functions are buggy in SOFTSUSY <=
3.6.2.

The bug in the SOFTSUSY's expression is in the terms which are
proportional to a soft-squared mass and which are not proportional to
g3 at the same time.

I found that at least a part of the problem is due to a false
identification of tr(Ye*Yec*me) with tr(Yt*Ytc*mt):

Example:

Consider the 3L beta function for m_H1^2.  In allgennb.log there exist
(at 3-loop level) the three terms

  a1^2 * ( - 234/25*tr(Yt*Ytc*mt)
           - 3069/50*tr(Ye*Yec*me)
           + 27/50*tr(Ye*Yec*me)*k + ... )

In SOFTSUSY's implementation, however, the terms involving
tr(Ye*Yec*me) are missing.  Instead, in SOFTSUSY one finds

  a1^2 * ( - 234/25*tr(Yt*Ytc*mt)
           - 3069/50*tr(Yt*Ytc*me)
           + 27/50*tr(Yt*Ytc*mt)*k + ... )