Bugfix: workaround inconsistently filled SARAH`Masses[EWSB]
for massless particles.

In the MDM SARAH`Masses[EWSB] reads

  {0 -> MassGiven[VG], Mass[gG] -> MassGiven[gG],
   Mass[Hp] -> MassGiven[Hp], Mass[Fv[1]] -> MassGiven[Fv[1]],
   Mass[Fv[2]] -> MassGiven[Fv[2]], Mass[Fv[3]] -> MassGiven[Fv[3]],
   0 -> MassGiven[Fc], 0 -> MassGiven[Fg], Mass[Fn] -> MassGiven[Fn],
   0 -> MassGiven[Fc], 0 -> MassGiven[Fg], Mass[Ah] -> MassGiven[Ah],
   ... }

Formerly, we've expected SARAH`Masses[EWSB] to contain replacement
rules of the form

  {Mass[X] -> _, ...}

Due to this expectation, some particles have not been extracted from
this list.

This commit works around this issue by extracting all particles X from
the list elements which are of the form

  Mass[X] -> _     or      _ -> MassGiven[X]