flag bad mass only for massive particles
The criterion for an imprecise mass output will almost always fail for
multiplets which contain at least one massless particle.  The reason
is, that the smallest eigenvalue will be close to zero, but not quite.
The criterion to flag a bad mass,

   eigenvalue_error > precision * Abs(MVPVZ(0))

will certainly be true, because on the r.h.s. you multiply my
approximately zero (the massless eigenvalue).