Use low scale constraint to set model parameters from SM parameters
This is, however, a very bad hack to achieve this, and needs to be
refined.  In particular, for consistency with SPheno, only
parameters depending on the gauge and Yukawa couplings should be
updated.  As part of this the ordering must be done correctly
to ensure e.g. the correct on-shell VEV is set.  Temporary
settings and macros must also be handled.