fill matching class only if FlexibleEFTHiggs === True
This avoids problems in models which are complicated to be matched to
the SM, as the munuSSM for example.  In the munuSSM it is not clear
how to match MFe to the SM.  Technically, the InvertMassRelation[]
function fails to solve for Ye for the matching class.  In order to
still be able to build a full model spectrum generator we don't
generate the matching class if FlexibleEFTHiggs =!= True.