clear pole masses after applying the matching
If an exception is thrown and the output is forced, the pole masses of
the SM particles in the full model, calculated at the matching scale,
are printed.  This can be irritating, because these masses are not the
final ones (and are probably very incorrect, because they have been
calculated at the wrong scale).