generalisation: allow usage of arbitrary CP-even Higgs mass entry
in BSM model.  For example in the NMSSM, one can now use the 2nd Higgs
to be the SM-like one.  However, this makes only sense, if the SM can
still be regarded as the correct EFT, i.e. when

   hh(0) ~ hh(1) ~ M_SUSY

In this case one would integrate out all particles at M_SUSY, which is
close to the EW scale.