Bugfix: Avoid conflict with Mathematica's internal MKL routines
Since version 9, Mathematica automatically opens its internal MKL
libraries.  It seems that when using FlexibleSUSY's LibraryLink
interface, the zgesvd_ function from MKL gets called, instead of the
one from the external LAPACK library.  This leads to a run-time
problem, because MKL's zgesvd_ has a different interface.  The error
messages were:

MKL ERROR: Parameter 4 was incorrect on entry to ZGESVD
MKL ERROR: Parameter 13 was incorrect on entry to ZGESVD

To fix the issue, we use Eigen's complex SVD routine when LibraryLink
is enabled.

This problem came up in the CMSSMCPV, where there is a complex 4x4
neutralino mass matrix, which needs to be diagonalized using SVD.