adding explicit dependency on .inc files to make them be generated
Note: The deeper problem is, that the .inc files don't yet exist and
need to be generated.  Still, they should appear in the list of the
dependencies, which can be triggered with the -MG option.

However, using the -MG option leads to a compiler error in the GM2Calc
interface, because our depgen.x does never ignore #include statements
wrapped in an #ifdef #endif block.  This needs to be fixed at some
point.