Bugfix: correcting SM 2-loop self energy O(alpha_t^2)
Thanks to Thomas Kwasnitza.

The old expression of Eq. (20) of arxiv:1205.6497 contains 2L
contributions from the momentum iteration of the 1-loop self energy.

However, FS does the momentum iteration by default.  For this reason,
these 2-loop pieces from the 1-loop momentum iteration must not be
included in the 2-loop that we add.

The expression of Eq. (20) of arxiv:1504.05200 contains only the self
energy contribution, but without the 2-loop contribution from the
1-loop momentum iteration.  This is therefore the 2-loop expression
that we should add to the (momentum-iterated) 1-loop expression.

The effect is of the order 50 MeV.