relax test pecision on fulfilled EWSB conditions
This relaxation is necessary for the new EWSB solving mechanism, where
the EWSB conditions are minimized by the solvers running on a *copy*
of the model object.

The cause of the slightly different EWSB solution is, that the
tree-level mass spectrum is calculated after the EWSB output
parameters have been set.  In other words: The old EWSB algorithm
(where the solvers do not run on a *copy* of the object, but on the
object itself, using the this pointer) leads to the same modified EWSB
solution if one calculates the tree-level spectrum right after the
EWSB output parameters have been set (after solve() has been called).

The changes are minimal if one keeps in mind that the EWSB equations
are of the dimension GeV^3.