use perfect forwarning in run_async() even for g++ 4.7.4
Note: g++ 4.7.4 does not support template parameter packs.  For this
reason, an overload of run_async() with 2 parameters is provided.

This commit introduces lambda expressions for the calculation of the
pole masses.  (The lambdas are perfectly forwarded to std::async().)
The template parameter type is now deduced by the compiler.