Bugfix: Use running precision 1 order smaller than precision goal
Peter found that the parameter point from below does not converge
during the determination of the SM(5) parameters.  The reason seems to
be that the used precision of the RG running (1e-5) is too small for
the parameters to converge (the reached precision for this point is
1.4e-5).

This commit fixes the issue by setting the running precision to 1
order of magnitude smaller than the precision goal of convergence.
This shold avoid such numerical instabilities.