New flag to disable calculation of non-SM pole masses.
This flag is useful in FlexibleEFTHiggs, when the SUSY scale is so
high that the non-SM particle masses become unreliable or tachyonic:
If a non-SM pole mass becomes tachyonic (maybe because the loop
corrections become too large) FlexibleSUSY would flag the given
parameter point as unphysical.  However, one might still be interested
in the value of the SM-like Higgs mass, which is valid in
FlexibleEFTHiggs even for very large SUSY scales.  In such a case
FlexibleSUSY[23] could be set to 0 to suppress the calculation of the
non-SM pole masses.