Bugfix: reset W pole mass in SM low-scale constraint
There are 2 different QedQcd objects: One in the Standard_model class,
and one in the SM low-scale constraint.  The W is re-calculated in the
Standard_model object.  Therefore, it must be reset in the low-scale
constraint as well.

This bug lead to wrong sin(theta_W) in scenarios with increased
alpha_em and/or decreased MZ.