remove QedQcd member from Standard_model class
as it is used only for initialization.

This change avoids bugs like the ones fixed in commit 20f169f by
design.  In addition the Standard_model class is now a bit smaller.

In addition the calculation of the Standard_model effective couplings
was optimized as in commit b533d67305c1b5aee7401dc136e9205c7fcb612d .