optimize convergence tester for FlexibleSUSY
The two-scale solver checks for convergence after all sliders have
been run.  In FlexibleSUSY the last slider stops at the SUSY scale.
This is the default scale where we want to calculate the spectrum and
thus the scale where we want to check for convergence.  Thus, by
default we don't need to specify a specific scale where to check for
convergence.

If we still do it, we decrease performance, because the SUSY scale
changes from iteration to iteration (if it is a dynmamic scale).
Thus, in every iteration we need to run all parameters to the new SUSY
scale inside the convergence tester (which has changed a bit).  This
leads to a performance drop.